feat: added access control for CRUD action for role section

+++ src/pages/Roles/index.tsx
@@ -13,10 +13,15 @@ import { convertQueryToString } from '@/utils/helpers/string';
 import { createLocale } from '@/config/translation/i18n';
 import { useDeleteRoleMutation, useGetRolesQuery } from '@/redux/apiSlice/rolesSlice';
 import { Role } from '@/@types/role.type';
+import { useAbility } from '@casl/react';
+import { AbilityContext } from '@/utils/providers/CanAbilityProvider';
+import { PermissionActions, PermissionModules } from '@/@types/permission.type';
 
 const Roles: React.FC = () => {
   const [isOpen, setIsOpen] = useState<boolean>(false);
   const [selected, setSelected] = useState<string | null>(null);
+
+  const ability = useAbility(AbilityContext);
   const { t } = useTranslation();
 
   const confirm = useConfirm();
@@ -38,28 +43,48 @@ const Roles: React.FC = () => {
     }
   }, []);
 
-  const actions = useMemo(() => [
-    {
-      id: 'edit',
-      label: t('general.edit'),
-      icon: <Edit className='w-4 h-4' />,
-      onClick: (res: Role) => {
-        setSelected(res.id);
-        setIsOpen(true);
-      }
-    },
-    {
-      id: 'delete',
-      label: t('general.delete'),
-      icon: <Delete className='w-4 h-4' />,
-      onClick: (res: Role) => handleDeleteRole(res)
-    },
-  ], []);
+  const actions = useMemo(() => {
+    const actionList = [];
+    if (ability.can(PermissionActions.Update, PermissionModules.Role)) {
+      actionList.push({
+        id: 'edit',
+        label: t('general.edit'),
+        icon: <Edit className='w-4 h-4' />,
+        onClick: (res: Role) => {
+          setSelected(res.id);
+          setIsOpen(true);
+        }
+      });
+    }
+    if (ability.can(PermissionActions.Delete, PermissionModules.Role)) {
+      actionList.push({
+        id: 'delete',
+        label: t('general.delete'),
+        icon: <Delete className='w-4 h-4' />,
+        onClick: (res: Role) => handleDeleteRole(res)
+      });
+    }
+    return actionList;
+  }, [ability]);
 
   const handleClose = useCallback(() => {
     setSelected(null);
     setIsOpen(false);
   }, []);
+
+  const extraControl = useMemo(() => {
+    if (ability.can(PermissionActions.Create, PermissionModules.Role)) {
+      return [
+        {
+          id: "add",
+          label: t("general.add"),
+          icon: <Add />,
+          onClick: () => setIsOpen(true)
+        }
+      ];
+    }
+    return [];
+  }, [ability]);
 
   return (
     <>
@@ -69,14 +94,7 @@ const Roles: React.FC = () => {
         data={data}
         head={createColumns(t)}
         actions={actions}
-        extraControl={[
-          {
-            id: "add",
-            label: t("general.add"),
-            icon: <Add />,
-            onClick: () => setIsOpen(true)
-          }
-        ]}
+        extraControl={extraControl}
       />
       <RoleFormModal open={isOpen} id={selected} onClose={handleClose} />
     </>
